* thread_pthread.c (ruby_init_stack): range of rlim_cur may be   larger than int.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@22898 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ thread_pthread.c
@@ -292,10 +292,10 @@ ruby_init_stack(VALUE *addr
 	struct rlimit rlim;
 
 	if (getrlimit(RLIMIT_STACK, &rlim) == 0) {
-	    unsigned int space = rlim.rlim_cur/5;
+	    size_t space = (size_t)(rlim.rlim_cur/5);
 
 	    if (space > 1024*1024) space = 1024*1024;
-	    native_main_thread.stack_maxsize = rlim.rlim_cur - space;
+	    native_main_thread.stack_maxsize = (size_t)rlim.rlim_cur - space;
 	}
     }
 #endif
